Work update and site adjustments
Managed to figure out how to correctly do the layout on the Work page using framer-motion, thanks to help by Viktor from framer. Now I just have to see if I want to get the card to dynamically expand or to just act as cool links to indiviual pages.

Redid how the navigation stays centered so that it adjusts more consistently.

Redesigned the logo to flow better with the amount of white space.

TODO: make the footer center better on mobile, easy fix is to use the same thing I did on the header and have it center relative to the entire page, might flow bettter.

TODO: finish the about page before the work, should be very simple.