Work Page Optimizations
Optimized the work page to now align the squares very beautifully on mobile. Took a bit of wrangling with  CSS properties to figure out how to get it to play nice, had to adjust the breakpoints to make the layout center properly when the elements were set vertically at the xs size.