About Page Mobile Optmisation
Hell yeah, got the about page scaling nice and proper on mobile. There is the small issue of the avatar image not scaling, not a bug, just need to see how to get the avatar material-ui component to be responsive OR I need ot use another component in its place, whichever is less hassle.