Navigation Improvements
Removed that awful hover effect on the material-ui navigation buttons.

I also repositioned the link tags to wrap the button tags so that click anywhere on the button works instead of the just on the text, idk why I did that in the first palce. Also move the onClick action from the text to the button for resume as well.