About Page Alignment Fix
Fixed the aligment and centering issues by removing the margin left default value on the text and making it change at the exact value it needs to. Now the image centers nicely on mobile and when the page is reszied.

Only thing left to do on the About page is to get the darn image to scale properly.

Immediate task afterwords is to fix the footer issue.