Big Small Push
Refactored some existing code across various components.

A lot of changes had to do with fixing how things resized, in particular the invisible Box spacers.

Implemented a solution for making the about page avatar image resize using breakpoints and a % width value instead of a px value. It is a little jerky with the way the breakpoints are set up but it does majorly solve the existing issue of the large image being static, so this can be revisting later.

Now focus should be on the WORK page, may need to layout the project links vertically to allow the project pages to load within the document, or will have it redirect to a new page entirely which may be easier, maybe.