jovisidhu-dev work page draft
Did a  lot of playing around with how to organize an arrange elements for a given portfolio component.

Still need to nail down how to resize the container to fit the needs of the content. If I remove the border and set the overflow to override the container's bounds then it may work, but may cause problem with bottom margins, which can be further ignored by keep all portfolio content in the hostContainer. Still might runover onto the footer, will test out later.

Text boxes with images interspersed on teh left and right seems interesting, but just having elements stacking in a flex column might be most convenient.