Work Component Styling Revision
Caved in, and decided to remove the border as it didn't really match nor complement the rest of the visual elements, and would be hard to style with other components.

I also did a quick draft of the writing for the intro of the pokedex page just to see how the site flowed from one work component to another.

How the elements are arranged on mobile needs to be addressed, the work component icons should not be big block elements, maybe rectangluar elements that just focus on encapsulating the project text.

TODO:  add close button to the bottom of the work compoenents, which should make it easier on mobile to get to the top when a work component has been expanded.