Work Page Structuring
Filled out a lot of the pages on the work tab.

Learned how to scale fonts and images properly, main thing is to use vw units.

Fixed vertical content bound issue on the work element content cotainer by removing the redundant inner container and not declaring a height value at all and the page handles itself.